chore: move Bubble to components
